fix(test): fix testLoginForAdmin_DPoP and address PR #2988 review feedback
- ChromeCustomTabPageObject: replace backOutToLoginActivity() with
  waitForLoginScreen() in tapLoginForAdminsMenuItem override, and set
  isBrowserLoginEnabled=false before tapping the server row so reloadWebView()
  loads the in-app WebView instead of relaunching a Custom Tab.
  This fixes the MORE_OPTIONS_BUTTON not found error after W-23731759 made
  the login picker non-dismissable.
- LoginViewModel: add defensive Welcome Discovery guard to addDpopJktIfNeeded
  for belt-and-suspenders symmetry with the iOS twin (optional suggestion from
  review, non-blocking).
- ECALoginTests: @Ignore testECAJwt_ViaLoginPoolServer until loginPoolHost is
  provisioned in the CI ui_test_config.json (blocking issue from PR review).
